* does not use collections to sort destinations
* uses ArrayBlockingQueue as the main queue

The new bundler works for me, though, I haven't found any tests that I should adapt and I have encountered errors when running `mvn test`:
` [testng] An error occurred while instantiating class org.jgroups.tests.byteman.ServerTest: org.jgroups.util.Util.close(Ljava/net/ServerSocket;)V`
So I haven't ran these.