@belaban 2015-12-01
Taking a look. I like the idea of getting rid of the hashmap in favor of a simple array, but...
* First benchmarks indicate this is slightly slower than TransferQueueBundler. I used 4 {{UPerf}} instances on the same box
* The hard limit of {{512}} now possibly triggers a bundled send() sooner than {{max_bundle_size}}; if we send a lot of very small messages, this will be suboptimal. Let's see if we can find a solution to this...
* When the array size (512) has been reached, the current sender sends all bundled messages. This is similar to the {{sender-sends}} bundler type, but now you have a problem: {{sendBundledMessages()}} needs to be protected, as both the bundler thread and a caller's thread can call this method concurrently!

Which perf test did you use to get the double-digit (iirc?) perf increase?