@rvansa 2015-12-02
I have used JMH benchmark that simulates 5 or 10 producers and single (S)TQB consumer. The code is a bit different; I had to refactor it for JMH and testing and then back, though the logic should be the same.
Here is the benchmark and adapted TQB:
https://github.com/rvansa/benchmarks/blob/master/ispntest/perftest/src/main/java/org/jboss/perf/hibernate/BundlerBenchmark.java
https://github.com/rvansa/benchmarks/tree/master/ispntest/perftest/src/main/java/org/jgroups/protocols
I have to try out UPerf as well to see the difference.

The hard limit is a guestimate - that's obviously something that could be configurable, but there's already too many knobs in JGroups configuration so I haven't added another one :) My concern was that if that array was too long, multiple iterations through it would lower the performance. And I am not optimizing for ultra-short messages as it's rather a corner case for Infinispan. Though, further suggestions are welcome.

Wrt `sendBundledMessages`: I don't think you're right, the sender threads call only `send(Message)` which enqueue the message (and blocks without space in the queue), while the array is operated by the bundler thread only.