@belaban 2015-12-02
Ah, ok, I thought `addMessage()` was called by the adder, but obviously that not the case. I'll take another look at the hard-coded array length tomorrow (was out of office today).