@belaban 2015-12-03
Applied your changes with modifications (mainly style). 2 things to revisit:
* The hard coded 512 sized buffer
* Moving bundlers out of TP. I want to be able to (e.g.) define the fully qualified classname of a bundler to be used.