@rvansa 2015-12-03
@belaban Just FYI, I've ran two simple ISPN benchmarks (4 nodes, 50 threads executing exclusively puts or gets, TCP stack) and I got these data:

Before:
265,388 gets/s, 753 us MRT
96,564 puts/s, 2.06 ms MRT

After:
268,457 gets/s, 744 us MRT
97,647 puts/s, 2.04 ms MRT

It seems that there's also some CPU usage reduction, but generally this is a marginal difference in performance. I'll have to run JFRs yet to see how memory consumption has changed; though, the effect won't be linear given the speedup (and especially as ISPN generates way lot of garbage).