@belaban 2015-12-03
Is this IspnPerfTest, JMH or Radargun?
For now, `SimplifiedTransferQueue` is in, but I'm not advertizing it yet (docs), so we can perf-test it, so no hurries.